CI: Move DOSv2 directory ops tests to separate file
